more clean ups

+++ src_cpp/igemm_params.cpp
@@ -27,18 +27,18 @@ void device::MixPrecAHA::rATA_params_query(gemm_params* param, double epi, int32
   param->precC = acc_bits <= 24 ? Precision::FP32 : (acc_bits <= 53 ? Precision::FP64 : f128);
   param->C_elem_bytes = acc_bits <= 24 ? 4 : (acc_bits <= 53 ? 8 : 16);
 
-  constexpr uint64_t algn_i8 = uint64_t(64) * sizeof(int32_t) - 1;
-  uint64_t strideA = uint64_t(param->algnM) * uint64_t(N);
-  uint64_t strideC = uint64_t(param->algnN) * uint64_t(N);
+  constexpr int64_t algn_i8 = int64_t(64) * sizeof(int32_t) - int64_t(1);
+  int64_t strideA = int64_t(param->algnM) * int64_t(N);
+  int64_t strideC = int64_t(param->algnN) * int64_t(N);
 
-  param->acc_bytes = uint64_t(param->C_elem_bytes) * strideC;
-  param->i8_bytes = uint64_t(param->orderA) * strideA;
+  param->acc_bytes = int64_t(param->C_elem_bytes) * strideC;
+  param->i8_bytes = int64_t(param->orderA) * strideA;
   param->i8_bytes = (param->i8_bytes + algn_i8) & (~algn_i8);
-  param->exp_bytes = sizeof(int32_t) * uint64_t(param->algnN);
-  param->scratch_bytes = sizeof(int32_t) * uint64_t(param->orderA + 1) * strideC;
+  param->exp_bytes = sizeof(int32_t) * int64_t(param->algnN);
+  param->scratch_bytes = sizeof(int32_t) * int64_t(param->orderA + 1) * strideC;
 
-  uint64_t A_elem_bytes = precA == Precision::FP32 ? sizeof(float) : sizeof(double);
-  uint64_t spare_space = std::max(param->i8_bytes + param->scratch_bytes, A_elem_bytes * strideC);
+  int64_t A_elem_bytes = precA == Precision::FP32 ? sizeof(float) : sizeof(double);
+  int64_t spare_space = std::max(param->i8_bytes + param->scratch_bytes, A_elem_bytes * strideC);
   param->C_bytes = param->acc_bytes + param->exp_bytes + spare_space;
 }
 
@@ -57,17 +57,17 @@ void device::MixPrecAHA::cAHA_params_query(gemm_params* param, double epi, int32
   param->precC = acc_bits <= 24 ? Precision::FP32 : (acc_bits <= 53 ? Precision::FP64 : f128);
   param->C_elem_bytes = acc_bits <= 24 ? 8 : (acc_bits <= 53 ? 16 : 32);
 
-  constexpr uint64_t algn_i8 = uint64_t(64) * sizeof(int32_t) - 1;
-  uint64_t strideA = uint64_t(2) * uint64_t(param->algnM) * uint64_t(N);
-  uint64_t strideC = uint64_t(param->algnN) * uint64_t(N);
+  constexpr int64_t algn_i8 = int64_t(64) * sizeof(int32_t) - int64_t(1);
+  int64_t strideA = int64_t(2) * int64_t(param->algnM) * int64_t(N);
+  int64_t strideC = int64_t(param->algnN) * int64_t(N);
 
-  param->acc_bytes = uint64_t(param->C_elem_bytes) * strideC;
-  param->i8_bytes = uint64_t(param->orderA) * strideA;
+  param->acc_bytes = int64_t(param->C_elem_bytes) * strideC;
+  param->i8_bytes = int64_t(param->orderA) * strideA;
   param->i8_bytes = (param->i8_bytes + algn_i8) & (~algn_i8);
-  param->exp_bytes = sizeof(int32_t) * uint64_t(param->algnN);
-  param->scratch_bytes = sizeof(int32_t) * uint64_t(param->orderA + 1) * strideC;
+  param->exp_bytes = sizeof(int32_t) * int64_t(param->algnN);
+  param->scratch_bytes = sizeof(int32_t) * int64_t(param->orderA + 1) * strideC;
 
-  uint64_t spare_space = std::max(param->i8_bytes + param->scratch_bytes, param->acc_bytes);
+  int64_t spare_space = std::max(param->i8_bytes + param->scratch_bytes, param->acc_bytes);
   param->C_bytes = param->acc_bytes + param->exp_bytes + spare_space;
 }
